Only run downloader tests on windows and linux since macos CI runs on arm which builds are not available for
